Rename test functions consistently
