+++ spec/helpers/lifecycle/failing-destroy-hook-streaming-function.js
