+++ spec/helpers/lifecycle/failing-init-hook-streaming-function.js
