+++ spec/helpers/lifecycle/simple-lifecycle-streaming-function.js
