+++ spec/helpers/lifecycle/slow-destroy-hook-streaming-function.js
