+++ spec/helpers/lifecycle/slow-init-hook-streaming-function.js
